Extract helpers to prevent a circular dependency in the future.
